+++ packages/client/components/SpotlightResultsEmptyState.tsx
@@ -3,14 +3,14 @@ import styled from '@emotion/styled'
 import purpleLines from '../styles/theme/images/purpleLines.svg'
 import {PALETTE} from '../styles/paletteV3'
 
-const EmptyState = styled('div')({
+const EmptyState = styled('div')<{height: number | string}>(({height}) => ({
   display: 'flex',
   justifyContent: 'center',
   alignContent: 'center',
   flexWrap: 'wrap',
-  minHeight: 320,
-  height: '100%'
-})
+  minHeight: 160,
+  height
+}))
 
 const MessageWrapper = styled('div')({
   display: 'flex',
@@ -44,18 +44,25 @@ const Content = styled('div')({
   alignItems: 'center'
 })
 
-const SpotlightGroupsEmptyState = () => (
-  <EmptyState>
-    <Emoji>😔</Emoji>
-    <Content>
-      <Img src={purpleLines} />
-      <MessageWrapper>
-        <Message>No reflections match this card.</Message>
-        <Message>Try searching for specific keywords.</Message>
-      </MessageWrapper>
-      <Img isFlipped src={purpleLines} />
-    </Content>
-  </EmptyState>
-)
+interface Props {
+  height: string | number
+}
 
-export default SpotlightGroupsEmptyState
+const SpotlightResultsEmptyState = (props: Props) => {
+  const {height} = props
+  return (
+    <EmptyState height={height}>
+      <Emoji>😔</Emoji>
+      <Content>
+        <Img src={purpleLines} />
+        <MessageWrapper>
+          <Message>No reflections match this card.</Message>
+          <Message>Try searching for specific keywords.</Message>
+        </MessageWrapper>
+        <Img isFlipped src={purpleLines} />
+      </Content>
+    </EmptyState>
+  )
+}
+
+export default SpotlightResultsEmptyState