+++ packages/client/hooks/useGroupMatrix.ts
@@ -1,11 +1,11 @@
-import {SpotlightGroupsQueryResponse} from './../__generated__/SpotlightGroupsQuery.graphql'
+import {SpotlightResultsQueryResponse} from './../__generated__/SpotlightResultsQuery.graphql'
 import {RefObject, useEffect, useState} from 'react'
 import {ElementWidth} from '../types/constEnums'
 import useResizeObserver from './useResizeObserver'
 import {MAX_SPOTLIGHT_COLUMNS} from '~/utils/constants'
 import useInitialRender from './useInitialRender'
 
-type Group = SpotlightGroupsQueryResponse['viewer']['similarReflectionGroups'][0]
+type Group = SpotlightResultsQueryResponse['viewer']['similarReflectionGroups'][0]
 
 const useGroupMatrix = (
   resultsGroups: readonly Group[],