+++ packages/client/modules/outcomeCard/components/OutcomeCardFooter/TaskFooter.tsx
@@ -69,7 +69,7 @@ const TaskFooter = (props: Props) => {
   }
   const handleCompleted = (res, errors) => {
     onCompleted(res, errors)
-    const payload = res?.[Object.keys(res)[0]]
+    const payload = res?.[Object.keys(res)[0]!]
     const error = payload?.error?.message ?? errors?.[0]?.message ?? null
     setLocalTaskError(atmosphere, taskId, error)
   }