+++ packages/server/graphql/intranetSchema/mutations/hardDeleteUser.ts
@@ -78,22 +78,26 @@ const hardDeleteUser = {
       swapCreatedByUserUpdates,
       discussions
     ] = await Promise.all([
-      (r
-        .table('MeetingMember')
-        .getAll(r.args(meetingIds), {index: 'meetingId'})
-        .group('meetingId') as any)
+      (
+        r
+          .table('MeetingMember')
+          .getAll(r.args(meetingIds), {index: 'meetingId'})
+          .group('meetingId') as any
+      )
         .count()
         .ungroup()
         .filter((row) => row('reduction').le(1))
         .map((row) => row('group'))
         .coerceTo('array')
         .run(),
-      (r
-        .table('NewMeeting')
-        .getAll(r.args(teamIds), {index: 'teamId'})
-        .filter((row) => row('meetingType').eq('retro'))
-        .eqJoin('id', r.table('RetroReflection'), {index: 'meetingId'})
-        .zip() as any)
+      (
+        r
+          .table('NewMeeting')
+          .getAll(r.args(teamIds), {index: 'teamId'})
+          .filter((row) => row('meetingType').eq('retro'))
+          .eqJoin('id', r.table('RetroReflection'), {index: 'meetingId'})
+          .zip() as any
+      )
         .filter((row) => row('creatorId').eq(userIdToDelete))
         .getField('id')
         .coerceTo('array')
@@ -140,18 +144,9 @@ const hardDeleteUser = {
 
     // all other writes
     await r({
-      teamMember: r
-        .table('TeamMember')
-        .getAll(userIdToDelete, {index: 'userId'})
-        .delete(),
-      meetingMember: r
-        .table('MeetingMember')
-        .getAll(userIdToDelete, {index: 'userId'})
-        .delete(),
-      notification: r
-        .table('Notification')
-        .getAll(userIdToDelete, {index: 'userId'})
-        .delete(),
+      teamMember: r.table('TeamMember').getAll(userIdToDelete, {index: 'userId'}).delete(),
+      meetingMember: r.table('MeetingMember').getAll(userIdToDelete, {index: 'userId'}).delete(),
+      notification: r.table('Notification').getAll(userIdToDelete, {index: 'userId'}).delete(),
       organizationUser: r
         .table('OrganizationUser')
         .getAll(userIdToDelete, {index: 'userId'})
@@ -176,10 +171,7 @@ const hardDeleteUser = {
         .getAll(r.args(teamIds), {index: 'teamId'})
         .filter((row) => r(teamMemberIds).contains(row('teamMemberId')))
         .delete(),
-      pushInvitation: r
-        .table('PushInvitation')
-        .getAll(userIdToDelete, {index: 'userId'})
-        .delete(),
+      pushInvitation: r.table('PushInvitation').getAll(userIdToDelete, {index: 'userId'}).delete(),
       retroReflection: r
         .table('RetroReflection')
         .getAll(r.args(retroReflectionIds), {index: 'id'})
@@ -215,7 +207,7 @@ const hardDeleteUser = {
           .table('NewMeeting')
           .get(update('id'))
           .update({
-            facilitatorUserId: (update('otherTeamMember') as unknown) as string
+            facilitatorUserId: update('otherTeamMember') as unknown as string
           })
       ),
       swapCreatedByUser: r(swapCreatedByUserUpdates).forEach((update) =>
@@ -223,7 +215,7 @@ const hardDeleteUser = {
           .table('NewMeeting')
           .get(update('id'))
           .update({
-            createdBy: (update('otherTeamMember') as unknown) as string
+            createdBy: update('otherTeamMember') as unknown as string
           })
       )
     }).run()