+++ packages/server/graphql/mutations/createGitHubTaskIntegration.ts
@@ -45,10 +45,7 @@ export default {
     const viewerId = getUserId(authToken)
 
     // AUTH
-    const task = await r
-      .table('Task')
-      .get(taskId)
-      .run()
+    const task = await r.table('Task').get(taskId).run()
     if (!task) {
       return standardError(new Error('Task not found'), {userId: viewerId})
     }
@@ -79,6 +76,9 @@ export default {
       dataLoader.get('teamMembersByTeamId').load(teamId)
     ])
     const auth = viewerAuth ?? assigneeAuth
+    if (!auth) {
+      return standardError(new Error('No auth exists for a given task!'), {userId: viewerId})
+    }
     const accessUserId = viewerAuth ? viewerId : assigneeAuth ? userId : null
     if (!accessUserId) {
       return standardError(
@@ -87,8 +87,15 @@ export default {
       )
     }
     // using teamMembers to get the preferredName as we need the members for the notification part anyways
-    const {preferredName: viewerName} = teamMembers.find(({userId}) => userId === viewerId)
-    const {preferredName: assigneeName} =
+    const teamMember = teamMembers.find(({userId}) => userId === viewerId)
+    if (!teamMember) {
+      return standardError(new Error('User is not member of the team'), {
+        userId: viewerId,
+        tags: {teamId}
+      })
+    }
+    const {preferredName: viewerName} = teamMember
+    const {preferredName: assigneeName = ''} =
       (userId && teamMembers.find((user) => user.userId === userId)) || {}
 
     // RESOLUTION