+++ packages/server/graphql/types/EstimateStage.ts
@@ -62,7 +62,7 @@ const EstimateStage = new GraphQLObjectType<Source, GQLContext>({
         const {service} = integration
         const getDimensionName = async (meetingId: string) => {
           const meeting = await dataLoader.get('newMeetings').load(meetingId)
-          const {templateRefId} = meeting
+          const {templateRefId} = meeting as MeetingPoker
           const templateRef = await dataLoader.get('templateRefs').load(templateRefId)
           const {dimensions} = templateRef
           const dimensionRef = dimensions[dimensionRefIdx]
@@ -146,7 +146,7 @@ const EstimateStage = new GraphQLObjectType<Source, GQLContext>({
           dataLoader.get('newMeetings').load(meetingId),
           dataLoader.get('meetingTaskEstimates').load({taskId, meetingId})
         ])
-        const {templateRefId} = meeting
+        const {templateRefId} = meeting as MeetingPoker
         const templateRef = await dataLoader.get('templateRefs').load(templateRefId)
         const {dimensions} = templateRef
         const dimensionRef = dimensions[dimensionRefIdx]
@@ -171,8 +171,7 @@ const EstimateStage = new GraphQLObjectType<Source, GQLContext>({
         const redis = getRedis()
         const userIds = await redis.smembers(`pokerHover:${stageId}`)
         if (userIds.length === 0) return []
-        const users = (await dataLoader.get('users').load(userIds)).filter(isValid)
-        return users
+        return (await dataLoader.get('users').loadMany(userIds)).filter(isValid)
       }
     },
     scores: {