+++ packages/server/postgres/queries/removeMattermostAuth.ts
@@ -1,0 +1,7 @@
+import getPg from '../getPg'
+import {removeMattermostAuthQuery} from './generated/removeMattermostAuthQuery'
+
+const removeMattermostAuth = async (userId: string, teamId: string) => {
+  await removeMattermostAuthQuery.run({userId, teamId}, getPg())
+}
+export default removeMattermostAuth